Add filters for ensuring deterministic snapshots (#32)
## Summary

<!-- Describe your changes in detail here, if it closes an open issue,
include "Closes #<issue>" -->
This PR adds the concept of "filters" to snapshot.

A filter is a regex replacement that is applied to snapshots prior to
serialisation, the goal of which is to replace non-deterministic (but
known format) data
with deterministic placeholders.

Some motivating examples:

- Any valid UUID -> `"[UUID]"`
- A unix timestamp -> `"[TIMESTAMP]"`
- Temp folders on various platforms -> `"[TEMP_FILE]"`
- etc.

I wonder if we should provide a set of defaults? And possibly a way to
opt out if not desired. Or just leave it completely up to the caller and
provide no defaults, but document
some examples (like the above)